dashboard/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
structlog>=24.0.0
pydantic>=2.0.0
websockets>=12.0
sse-starlette>=2.0.0
aiofiles>=23.0.0
